% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_inews.R
\name{.onLoad}
\alias{.onLoad}
\title{Simple preloaded functions
Updates ggplot defaults +  imports relevant fonts to registry + sets helpful global variables}
\usage{
.onLoad(libname, pkgname)
}
\description{
Simple preloaded functions
Updates ggplot defaults +  imports relevant fonts to registry + sets helpful global variables
}
