% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_inews.R
\name{get_inews_scales}
\alias{get_inews_scales}
\title{Collection of colour scales for Inews}
\usage{
get_inews_scales(palette, n, direction)
}
\arguments{
\item{palette}{palette to go to get_inews_scale}

\item{n}{number of colours to return}

\item{direction}{whether to reverse palette}
}
\description{
Collection of colour scales for Inews
}
