% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_inews.R
\name{guide_coloursteps_inews}
\alias{guide_coloursteps_inews}
\title{Internal function to enable raster on coloursteps, seemingly not possible natively in ggplot2}
\usage{
guide_coloursteps_inews(
  even.steps = TRUE,
  show.limits = NULL,
  ticks = FALSE,
  ...
)
}
\description{
Internal function to enable raster on coloursteps, seemingly not possible natively in ggplot2
}
