% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_inews.R
\name{inews_fill_steps}
\alias{inews_fill_steps}
\title{Generate binned gradients by specifying bin width}
\usage{
inews_fill_steps(
  ...,
  bvals = NULL,
  pallette = "blue",
  width = NULL,
  values = NULL,
  colours = NULL,
  breaks = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "coloursteps",
  aesthetics = "fill"
)
}
\arguments{
\item{bvals}{Data to split into breaks}

\item{pallette}{Prespecified pallettes}

\item{width}{Size of each bin}
}
\description{
Generate binned gradients by specifying bin width
}
