% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_inews.R
\name{inews_pal}
\alias{inews_pal}
\title{Inews Pallette}
\usage{
inews_pal(palette = "qual", length = NA, direction = 1)
}
\arguments{
\item{palette}{A collection of different palettes, along with discrete palette to be used easily with scale_inews_ferm()}

\item{length}{the number of colours for the palette, passed automatically from break length}

\item{direction}{The direction of palette, passed from scale_inews_ferm}
}
\description{
Inews Pallette
}
