% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_inews.R
\name{labels_inews}
\alias{labels_inews}
\title{Inews labelling}
\usage{
labels_inews(label, size = 0.6, type = "line")
}
\arguments{
\item{label}{passes aes to directlabels to allow for quicker line labelling}

\item{size}{allows to increase size of labels}

\item{type}{Horizontal labels for bar charts, vertical for lines}
}
\description{
Inews labelling
}
