% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_inews.R
\name{scale_inews_ferm}
\alias{scale_inews_ferm}
\title{Binned scale (scale_fill_fermenter wrapper with better gradients)}
\usage{
scale_inews_ferm(
  breaks,
  db = F,
  n = 5,
  direction = 1,
  style = "pretty",
  palette = "red_blue",
  na.value = "grey50",
  guide = "coloursteps",
  ...
)
}
\arguments{
\item{n}{number of breaks (if not direct)}

\item{direction}{whether to reverse palette}

\item{style}{to be passed on to classInt}

\item{palette}{palette to go to get_inews_scale}
}
\description{
Binned scale (scale_fill_fermenter wrapper with better gradients)
}
