% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_inews.R
\name{theme_inews_map}
\alias{theme_inews_map}
\title{Inews map theme}
\usage{
theme_inews_map(base_size = 25, base_family = "", fill = "White")
}
\arguments{
\item{base_size}{Basic size of graph, defaults to 25}

\item{base_family}{base font family, not implemented}
}
\description{
Inews map theme
}
