% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_inews.R
\name{save_inews}
\alias{save_inews}
\title{Save plots (ggsave wrapper)}
\usage{
save_inews(
  filename,
  plot = last_plot(),
  width_i = 15,
  height_i = 10,
  type = "basic",
  l_size = TRUE
)
}
\arguments{
\item{filename}{The filename specified}

\item{plot}{The plot to render, defaults to last plot}

\item{width_i}{Specified width in cm, defaults to 15}

\item{height_i}{Specified height in cm, defaults to 10}

\item{type}{Adds presets for maps/parls and other types to render}

\item{l_size}{Enables to turn off limiting size}
}
\description{
Save plots (ggsave wrapper)
}
