% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_inews.R
\name{save_inews}
\alias{save_inews}
\title{Save plots (ggsave wrapper)}
\usage{
save_inews(
  filename,
  plot = last_plot(),
  width = 15,
  height = 11,
  type = "basic",
  units = "cm",
  margin = "none"
)
}
\arguments{
\item{filename}{The filename specified}

\item{plot}{The plot to render, defaults to last plot}

\item{width}{Specified width in cm, defaults to 15}

\item{height}{Specified height in cm, defaults to 11}

\item{type}{Adds presets for different graph types or different places on website}

\item{units}{Specift units for width and height}

\item{margin}{Add margins to plot}
}
\description{
Save plots (ggsave wrapper)
}
