% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_inews.R
\name{scale_inews_ferm}
\alias{scale_inews_ferm}
\title{Fermenter scale for maps}
\usage{
scale_inews_ferm(
  palette = palette,
  breaks = breaks,
  direction = 1,
  labels = waiver(),
  na.value = "#DCDCDC",
  ...
)
}
\arguments{
\item{palette}{palette string from inews_pal}

\item{direction}{direction of palette, passed to inews_pal}

\item{labels}{add labels if % etc.}

\item{na.value}{Na value, defaults to light grey}

\item{break}{must be specified, determines when to break map distitions}
}
\description{
Fermenter scale for maps
}
